+++ NEWS.md
@@ -4,6 +4,8 @@
 
 ## Enhancements
 
+* Make local file path more robust in test-suite
+* Fix non-monotonicity bugs for grouped distributions
 * Remove `purrr` dependency
 * add test for `sd_create_synth_vector`
 * create a separate function (`weighted_average_poverty_stats`) for repetitive calculation in adjust poverty stats and add corresponding test case for it.